+++ R/norm_BOLD.R
@@ -4,11 +4,15 @@
 #'  For dual regression, row centering is required and column centering is not
 #'  recommended. Scaling and detrending depend on the user preference.
 #'
+#' @template superseded-by-BayesBrainMap
 #' @param BOLD fMRI numeric data matrix (\eqn{V \times T})
 #' @param center_rows,center_cols Center BOLD data across rows (each data
 #'  location's time series) or columns (each time point's image)? Default:
 #'  \code{TRUE} for row centering, and \code{FALSE} for column centering.
-#' @inheritParams scale_Param
+#' @param scale \code{"global"} (default), \code{"local"}, or \code{"none"}.
+#'  Global scaling will divide the entire data matrix by the mean image standard
+#'  deviation (\code{mean(sqrt(rowVars(BOLD)))}). Local scaling will divide each
+#'  data location's time series by its estimated standard deviation.
 #' @param scale_sm_xifti,scale_sm_FWHM Only applies if \code{scale=="local"} and
 #'  \code{BOLD} represents CIFTI-format data. To smooth the standard deviation
 #'  estimates used for local scaling, provide a \code{"xifti"} object with data
@@ -18,20 +22,32 @@
 #' @param scale_sm_xifti_mask For local scaling with smoothing, the data must
 #'  be unmasked to be mapped back to the surface. So if the data are masked,
 #'  provide the mask here.
-#' @inheritParams TR_param
-#' @inheritParams hpf_param
+#' @param TR The temporal resolution of the data, i.e. the time between volumes,
+#'  in seconds. \code{TR} is required for detrending with \code{hpf}.
+#' @param hpf,lpf The frequencies at which to apply temporal filtering to the
+#'  data during pre-processing, in Hertz. Set either to \code{NULL} to disable.
+#'  Default: \code{0.01} Hz highpass filter, and \code{NULL} for the lowpass 
+#'  filter (disabled). Filtering is accomplished by nuisance regression of
+#'  discrete cosine transform (DCT) bases.
+#' 
+#'  The highpass filter serves to detrend the data, since low-frequency 
+#'  variance is associated with noise. The lowpass filter removes high-frequency
+#'  variance, which is also thought to be from non-neuronal noise.
+#' 
+#'  Note the \code{TR} argument is required for temporal filtering. If
+#'  \code{TR} is not provided, \code{hpf} and \code{lpf} will be ignored.
 #'
 #' @return Normalized BOLD data matrix (\eqn{V \times T})
 #'
 #' @export
 #'
-#' @importFrom fMRItools nuisance_regression dct_bases dct_convert
+#' @importFrom fMRItools nuisance_regression temporal_filter
 #'
 norm_BOLD <- function(
   BOLD, center_rows=TRUE, center_cols=FALSE,
   scale=c("local", "global", "none"), scale_sm_xifti=NULL, scale_sm_FWHM=2,
   scale_sm_xifti_mask=NULL,
-  TR=NULL, hpf=.01){
+  TR=NULL, hpf=.01, lpf=NULL){
 
   nT <- ncol(BOLD)
   nV <- nrow(BOLD)
@@ -59,17 +75,24 @@ norm_BOLD <- function(
     }
   }
   stopifnot(is.numeric(scale_sm_FWHM) && length(scale_sm_FWHM)==1)
-  if (is.null(hpf)) { hpf <- 0 }
+  if (length(hpf)==1 && hpf==0) { hpf <- NULL }
+  if (length(lpf)==1 && lpf==Inf) { lpf <- NULL }
   if (is.null(TR)) {
-    if (hpf==.01) {
-      message("Setting `hpf=0` because `TR` was not provided. Either provide `TR` or set `hpf=0` to disable this message.")
-      hpf <- 0
-    } else if (hpf!=0) {
-      stop("Cannot apply `hpf` because `TR` was not provided. Either provide `TR` or set `hpf=0`.")
+    if (!is.null(hpf)) {
+      if (hpf==.01) {
+        message("Setting `hpf=NULL` because `TR` was not provided. Either provide `TR` or set `hpf=NULL` to disable this message.")
+        hpf <- NULL
+      } else {
+        stop("Cannot apply `hpf` because `TR` was not provided. Either provide `TR` or set `hpf=NULL`.")
+      }
+    }
+    if (!is.null(lpf)) {
+      stop("Cannot apply `lpf` because `TR` was not provided. Either provide `TR` or set `lpf=NULL`.")
     }
   } else {
-    stopifnot(fMRItools::is_posNum(TR))
-    stopifnot(fMRItools::is_posNum(hpf, zero_ok=TRUE))
+    stopifnot(is_posNum(TR))
+    stopifnot(is.null(hpf) || is_posNum(hpf))
+    stopifnot(is.null(lpf) || is_posNum(lpf))
   }
 
   # Center.
@@ -89,19 +112,14 @@ norm_BOLD <- function(
     }
   }
 
-  # Detrend.
-  # [NOTE]: If `center_cols`, columns won't be centered anymore after detrending.
-  if (hpf > 0) {
-    nDCT <- round(fMRItools::dct_convert(T_=nT, TR=TR, f=hpf))
-    if (nDCT == 0) {
-      warning("For the low `hpf` and at the data TR and length, the closest number of DCT bases to use for detrending is zero. Using one instead. See `fMRItools::dct_convert`.")
-      nDCT <- 1
-    }
+  # Apply the temporal filter.
+  # [NOTE]: If `center_cols`, columns won't be exactly centered anymore after the filter.
+  if (!is.null(hpf) || !is.null(lpf)) {
     if (!center_rows) { voxMeans <- rowMeans(BOLD, na.rm=TRUE) }
-    BOLD <- fMRItools::nuisance_regression(
-      BOLD,
-      cbind(1, fMRItools::dct_bases(nT, nDCT))
-    )
+    dct <- fMRItools::temporal_filter(
+      X=ncol(BOLD), TR=TR, hpf=hpf, lpf=lpf, method="DCT", verbose=FALSE
+    ) # [TO DO] carry over verbose arg?
+    BOLD <- nuisance_regression(BOLD, cbind(1, dct))
     if (!center_rows) { BOLD <- BOLD + voxMeans }
   }
 
@@ -137,7 +155,12 @@ norm_BOLD <- function(
       # Compute and smooth the SD.
       sig <- ciftiTools::newdata_xifti(ciftiTools::select_xifti(scale_sm_xifti, 1), sig)
       sig <- ciftiTools::move_to_mwall(sig, NA)
-      sig_mask <- do.call(c, sig$meta$cortex$medial_wall_mask)
+      if (!is.null(sig$data$subcort)) {
+        sub_mask <- !is.na(sig$data$subcort[,1])
+        sig$data$subcort <- sig$data$subcort[sub_mask,,drop=FALSE]
+        sig$meta$subcort$labels <- sig$meta$subcort$labels[sub_mask]
+        sig$meta$subcort$mask[sig$meta$subcort$mask][!sub_mask] <- FALSE
+      }
       sig <- ciftiTools::smooth_xifti(sig, surf_FWHM=scale_sm_FWHM, vol_FWHM=scale_sm_FWHM)
       sig <- c(as.matrix(sig))
     }